requirements.txt
aiohttp
fastapi
python-dotenv
uvicorn
